fix: resolve all 40 Codacy issues — split large files, fix security, reduce complexity
Split 6 oversized files (>500 NLOC) into focused modules:
- test_check_codacy_zero.py -> core + edge + retry (1266 -> 413+472+494)
- check_codacy_zero.py -> main + codacy_zero_helpers.py (715 -> 497+105)
- test_run_coverage_gate.py -> core + extra (664 -> 354+404)
- test_check_sonar_zero.py -> core + edge (614 -> 273+437)
- test_quality_common.py -> core + extra (569 -> 320+356)
- test_coverage_assert.py -> core + extra (567 -> 313+312)

Harden subprocess security in script_entrypoint_support.py:
- Add input validation for script paths
- Add explicit shell=False to subprocess.run
- Extract bootstrap lines to a module-level constant
